Turns out, there are more than 255 seconds in a year

+++ firmware/nRF51/reader-prox-microbit/src/clock.c
@@ -26,7 +26,7 @@
 #include <timer.h>
 #include <clock.h>
 
-static volatile uint8_t g_clock_time;
+static volatile uint32_t g_clock_time;
 
 void RTC0_IRQ_Handler(void)
 {
